Add enable option
Given the nature of hooks it can be awkward to skip them
programmatically if they don't support this option directly.

This PR addresses
https://github.com/JohannesKlauss/react-hotkeys-hook/issues/480

And adds an option to `enable` installation of the hook, this option
defaults to `true`.  When set to `false`, no hotkey is installed, this
avoids the issue where you try and do the same thing using a filter and
forget to set `filterPreventDefault` to `false`.

Note that this has only been tested in my consuming app since I can't
get the gatsby stuff working ATM.